// src/app/models/kyc/KycModal.ts
export class KycModal {
  userId: string
  addressFrontFileURL: string;
  addressBackFileURL: string;
  addressProofNumber: string;
  addressResult: string;
  addressProofType: string;
  panFileURL: string;
  panNumber: string;
  idResult: string;
  documentType: string;
  kycStatus: string;
  createdAt: string;
  updatedAt: string;
  idReason: string;
  addressReason: string
}